+++ pillow_heif/__init__.py
@@ -11,6 +11,7 @@ from ._lib_info import (
 )
 from ._options import options
 from ._version import __version__
+from .as_opener import HeifImageFile, register_heif_opener
 from .constants import heif_filetype_maybe  # DEPRECATED
 from .constants import heif_filetype_no  # DEPRECATED
 from .constants import heif_filetype_yes_supported  # DEPRECATED
@@ -28,7 +29,7 @@ from .constants import (
 from .error import HeifError
 
 # pylint: disable=redefined-builtin
-from .Heif import (
+from .heif import (
     HeifFile,
     HeifImage,
     HeifThumbnail,
@@ -41,5 +42,4 @@ from .Heif import (
     read,
     read_heif,
 )
-from .HeifImagePlugin import HeifImageFile, register_heif_opener
 from .misc import get_file_mimetype, reset_orientation